subo funcionalidad completa de mensajes al anunciante

+++ Database/CREATE_Inmobiliaria_DB.sql
@@ -103,11 +103,11 @@ CREATE TABLE Favorito (
 );
 
 Create Table Mensajes (
-	IdMensaje bigint primary key not null identity (1, 1),
+	IdMensaje int primary key not null identity (1, 1),
 	IdPropiedad int foreign key references Propiedad(IdPropiedad),
-	IdUsuario int foreign key references Usuario(IdUsuario),
+  IdUsuario int foreign key references Usuario(IdUsuario),
 	Mensaje Varchar(255) not null,
-	NombreUsuario Varchar(100) not null,
+  NombreUsuario Varchar(100) not null,
 	FechaDePublicacion datetime NOT NULL DEFAULT GETDATE()
 )
